+++ lightly_studio/src/lightly_studio/resolvers/sample_resolver/sample_filter.py
@@ -25,6 +25,16 @@ class SampleFilter(BaseModel):
     sample_ids: Optional[list[UUID]] = None
     has_captions: Optional[bool] = None
     annotations_filter: Optional[AnnotationsFilter] = None
+
+    # Query expression filter
+    #
+    # Adds an arbitrary "where" condition that can be expressed with QueryExpr.
+    #
+    # Important note: The expression can (and usually will) reference fields outside of SampleTable.
+    # It is caller's responsibility to ensure that the necessary joins are present in the query
+    # before applying this filter. Usually, this is done by using SampleFilter from another
+    # class, e.g. ImageFilter. That's also the reason why we test this filter outside of
+    # test_sample_filter.py.
     query_expr: Optional[QueryExpr] = None
 
     def apply(self, query: QueryType) -> QueryType:
