+++ lightly_studio/tests/resolvers/test_image_filter.py
@@ -9,13 +9,6 @@ import pytest
 from sqlmodel import Session, col, select
 
 from lightly_studio.models.image import ImageTable
-from lightly_studio.models.query_expr import (
-    AndExpr,
-    FieldRef,
-    IntegerExpr,
-    OrdinalComparisonOperator,
-    QueryExpr,
-)
 from lightly_studio.resolvers import tag_resolver
 from lightly_studio.resolvers.image_filter import (
     FilterDimensions,
@@ -254,105 +247,3 @@ class TestImageFilter:
         assert [sample.sample_id for sample in result] == [
             sample.sample_id for sample in expected_samples
         ]
-
-    def test_apply__query_expr_filter(self, db_session: Session) -> None:
-        collection = create_collection(session=db_session)
-        samples = create_images(
-            db_session=db_session,
-            collection_id=collection.collection_id,
-            images=[
-                ImageStub(path="small.png", width=100),
-                ImageStub(path="large.png", width=800),
-            ],
-        )
-
-        image_filter = ImageFilter(
-            sample_filter=SampleFilter(
-                query_expr=QueryExpr(
-                    match_expr=IntegerExpr(
-                        field=FieldRef(table="image", name="width"),
-                        operator=OrdinalComparisonOperator.GT,
-                        value=500,
-                    )
-                )
-            )
-        )
-
-        query = select(ImageTable).join(ImageTable.sample)
-        result = db_session.exec(image_filter.apply(query=query)).all()
-
-        assert len(result) == 1
-        assert result[0].sample_id == samples[1].sample_id
-
-    def test_apply__query_expr_with_and(self, db_session: Session) -> None:
-        collection = create_collection(session=db_session)
-        samples = create_images(
-            db_session=db_session,
-            collection_id=collection.collection_id,
-            images=[
-                ImageStub(path="a.png", width=800, height=600),
-                ImageStub(path="b.png", width=800, height=200),
-                ImageStub(path="c.png", width=200, height=200),
-            ],
-        )
-
-        image_filter = ImageFilter(
-            sample_filter=SampleFilter(
-                query_expr=QueryExpr(
-                    match_expr=AndExpr(
-                        children=[
-                            IntegerExpr(
-                                field=FieldRef(table="image", name="width"),
-                                operator=OrdinalComparisonOperator.GT,
-                                value=500,
-                            ),
-                            IntegerExpr(
-                                field=FieldRef(table="image", name="height"),
-                                operator=OrdinalComparisonOperator.LT,
-                                value=400,
-                            ),
-                        ]
-                    )
-                )
-            )
-        )
-
-        query = select(ImageTable).join(ImageTable.sample)
-        result = db_session.exec(image_filter.apply(query=query)).all()
-
-        assert len(result) == 1
-        assert result[0].sample_id == samples[1].sample_id
-
-    def test_apply__query_expr_combined_with_sample_ids(self, db_session: Session) -> None:
-        collection = create_collection(session=db_session)
-        samples = create_images(
-            db_session=db_session,
-            collection_id=collection.collection_id,
-            images=[
-                ImageStub(path="a.png", width=800),
-                ImageStub(path="b.png", width=800),
-                ImageStub(path="c.png", width=200),
-            ],
-        )
-
-        # query_expr matches samples[0] and samples[1] (width > 500)
-        # sample_ids restricts to samples[1] and samples[2]
-        # AND gives only samples[1]
-        image_filter = ImageFilter(
-            sample_filter=SampleFilter(
-                sample_ids=[samples[1].sample_id, samples[2].sample_id],
-                query_expr=QueryExpr(
-                    match_expr=IntegerExpr(
-                        field=FieldRef(table="image", name="width"),
-                        operator=OrdinalComparisonOperator.GT,
-                        value=500,
-                    )
-                ),
-            )
-        )
-
-        query = select(ImageTable).join(ImageTable.sample)
-        result = db_session.exec(image_filter.apply(query=query)).all()
-
-        assert len(result) == 1
-        assert result[0].sample_id == samples[1].sample_id
